# Jaeger/Zipkin Exporter Performance