--- a/.github/workflows/dotnet-core-linux.yml
+++ b/.github/workflows/dotnet-core-linux.yml
@@ -27,6 +27,11 @@ jobs:
       with:
         dotnet-version: 3.1.301
 
+    # For linux, we have to apply a workaround to enable both dotnet versions at same time	
+    - name: .net SxS	
+      run: |	
+        rsync -a ${DOTNET_ROOT/3.1.301/2.1.807}/* $DOTNET_ROOT/
+
     - name: Install dependencies
       run: dotnet restore