--- a/src/OpenTelemetry.Exporter.Jaeger/CHANGELOG.md
+++ b/src/OpenTelemetry.Exporter.Jaeger/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Renamed `MaxPacketSize` -> `MaxPayloadSizeInBytes` on `JaegerExporterOptions`.
+  Lowered the default value from 65,000 to 4096.
+  ([#1247](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1274))
+
 ## 0.6.0-beta.1
 
 Released 2020-Sep-15