--- a/src/OpenTelemetry.Exporter.Jaeger/JaegerExporterOptions.cs
+++ b/src/OpenTelemetry.Exporter.Jaeger/JaegerExporterOptions.cs
@@ -22,7 +22,7 @@ namespace OpenTelemetry.Exporter.Jaeger
     {
         internal const string DefaultServiceName = "OpenTelemetry Exporter";
 
-        internal const int DefaultMaxPacketSize = 4096;
+        internal const int DefaultMaxPayloadSizeInBytes = 4096;
 
         /// <summary>
         /// Gets or sets the name of the service reporting telemetry. Default value: OpenTelemetry Exporter.
@@ -40,9 +40,9 @@ namespace OpenTelemetry.Exporter.Jaeger
         public int AgentPort { get; set; } = 6831;
 
         /// <summary>
-        /// Gets or sets the maximum packet size in bytes. Default value: 4096.
+        /// Gets or sets the maximum payload size in bytes. Default value: 4096.
         /// </summary>
-        public int? MaxPacketSize { get; set; } = DefaultMaxPacketSize;
+        public int? MaxPayloadSizeInBytes { get; set; } = DefaultMaxPayloadSizeInBytes;
 
         /// <summary>
         /// Gets or sets the tags that should be sent with telemetry.