… src/OpenTelemetry.Exporter.Zipkin/AssemblyInfo.cs