## Changes

A round of performance tweaks for Zipkin. It now uses the `Batch<Activity` throughout (instead of making a copy) and flushes small (4k) chunks to the Http stream as they are ready.

The new version is slightly slower. If I make the chunk size bigger, it can be faster. But I think that is cheating for the sake of benchmarking. In a real-world scenario, we want to start sending bytes as soon as we can. Today we buffer everything into memory and then send it all at the end, that's why the memory usage is so high. I think lower chunks will actually make the export finish faster for real-world usage. I did the same thing (lower packet size) for JaegerExporter. Size reduced from 65000 to 4096. Socket default send buffer is [8192](https://docs.microsoft.com/en-us/dotnet/api/system.net.sockets.socket.sendbuffersize?view=netcore-3.1#property-value). For Jaeger we want to get the packets to the socket and let the networking algorithm handle the transmission, it has enough buffering we don't need our own :smile:

/cc @reyang Curious what your thoughts are on this 4096 size change?

## Benchmarks

Before:
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Completed Work Items | Lock Contentions |       Gen 0 |       Gen 1 |       Gen 2 |  Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|---------------------:|-----------------:|------------:|------------:|------------:|-----------:|
| ZipkinExporter_Batching |               1 |         10000 |    49.77 ms |  0.539 ms |  0.478 ms |               5.1818 |                - |   1181.8182 |   1090.9091 |   1000.0000 |    18.7 MB |
| ZipkinExporter_Batching |              10 |         10000 |   458.13 ms |  5.493 ms |  4.869 ms |              34.0000 |                - |  11000.0000 |   9000.0000 |   9000.0000 |  186.79 MB |
| ZipkinExporter_Batching |             100 |         10000 | 4,625.93 ms | 72.792 ms | 68.090 ms |             305.0000 |                - | 124000.0000 | 100000.0000 | 100000.0000 | 1867.75 MB |

After:
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Completed Work Items | Lock Contentions |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|---------------------:|-----------------:|-----------:|----------:|------:|----------:|
| ZipkinExporter_Batching |               1 |         10000 |    56.84 ms |  1.087 ms |  1.209 ms |               4.2000 |                - |   200.0000 |         - |     - |   1.71 MB |
| ZipkinExporter_Batching |              10 |         10000 |   511.93 ms |  5.001 ms |  4.678 ms |              25.0000 |                - |  2000.0000 |         - |     - |  16.95 MB |
| ZipkinExporter_Batching |             100 |         10000 | 5,151.66 ms | 66.415 ms | 58.875 ms |             204.0000 |                - | 21000.0000 | 1000.0000 |     - |  169.3 MB |

## TODOs

* [X] `CHANGELOG.md` updated for non-trivial changes
* [ ] Changes in public API reviewed